+++ niveshpy/models/transaction.py
@@ -30,6 +30,13 @@ class TransactionType(StrEnum):
     However, you may use it for other types of transactions as well.
     """
 
+    REVERSAL = auto()
+    """Transaction type representing reversals.
+
+    This type indicates a transaction that reverses a previous transaction.
+    It is used to correct errors or negate the effects of a prior transaction.
+    """
+
 
 _quantize_units = functools.partial(quantize_decimal, places=3)
 _quantize_amount = functools.partial(quantize_decimal, places=2)
@@ -49,6 +56,8 @@ class TransactionCreate:
         account_id (int): Foreign key to the associated account.
         properties (dict[str, Any], optional): Additional properties of the transaction.
             Defaults to an empty dictionary.
+        is_ignored (bool, optional): Flag indicating if the transaction should be ignored.
+            Defaults to False.
     """
 
     transaction_date: date
@@ -59,6 +68,7 @@ class TransactionCreate:
     security_key: str
     account_id: int
     properties: Mapping[str, Any] = field(factory=dict)
+    is_ignored: bool = field(default=False, kw_only=True)
 
 
 @frozen
@@ -76,6 +86,7 @@ class TransactionPublic:
         account_id (int): Foreign key to the associated account.
         properties (Mapping[str, Any]): Additional properties of the transaction.
         created (datetime): Timestamp when the transaction was created.
+        is_ignored (bool): Flag indicating if the transaction should be ignored.
         security (SecurityPublic | None): Related security object, if available.
         account (AccountPublic | None): Related account object, if available.
         cost (Decimal | None): Cost basis of the transaction, if applicable.
@@ -92,6 +103,7 @@ class TransactionPublic:
     account_id: int
     properties: Mapping[str, Any]
     created: datetime
+    is_ignored: bool = False
     cost: Decimal | None = None
     security: SecurityPublic | None = None
     account: AccountPublic | None = None